Pass the Header theme functionality to the <Header /> component Use React's Fragments instead of <div> wrappers
