+++ src/App.tsx
@@ -16,7 +16,6 @@ import Landing from './components/Landing/Landing';
 import NotFound from './components/NotFound/NotFound';
 
 interface IAppState {
-	headerStyle: string,
 	user?: object
 }
 
@@ -24,27 +23,10 @@ class App extends React.Component<{}, IAppState> {
 	public constructor(props: {}) {
 		super(props);
 		this.state = {
-			headerStyle: "dark",
 			user: undefined,
 		};
-		this.onScroll = this.onScroll.bind(this)
-	}
-	public onScroll() {
-		const elems = Array.prototype.slice.call(document.querySelectorAll('.header-change-light, .header-change-dark'));
-		const elem = elems.find((e: HTMLElement) => {
-			const bounding = e.getBoundingClientRect();
-			return (bounding.top + bounding.height) > 60
-		});
-		if (elem) {
-			const headerStyle = elem.classList.contains('header-change-light') ? 'light' : 'dark'
-			if (this.state.headerStyle !== headerStyle) {
-				this.setState({ headerStyle })
-			}
-		}
 	}
 	public componentDidMount() {
-		window.addEventListener('scroll', this.onScroll)
-		this.onScroll();
 		history.listen((location, action) => {
 			console.log('Change!') // tslint:disable-line
 			// requestAnimationFrame(() => this.onScroll());
@@ -52,20 +34,17 @@ class App extends React.Component<{}, IAppState> {
 	}
 	public render() {
 		return (
-			<div>
-				<Router history={history}>
-					<div>
-						<Header theme={this.state.headerStyle}/>
-						<Switch>
-							<Route exact={true} path="/" component={Landing} />
-							<Route path="/editor" component={Editor} />
-
-							<Route component={NotFound} />
-						</Switch>
-						<Footer />
-					</div>
-				</Router>
-			</div>
+			<Router history={history}>
+				<>
+					<Header />
+					<Switch>
+						<Route exact={true} path="/" component={Landing} />
+						<Route path="/editor" component={Editor} />
+						<Route component={NotFound} />
+					</Switch>
+					<Footer />
+				</>
+			</Router>
 		);
 	}
 }
