+++ src/components/Header/Header.tsx
@@ -5,17 +5,42 @@ import './Header.css';
 
 import SearchBox from './SearchBox/SearchBox';
 
-interface IHeaderProps {
+interface IHeaderState {
 	theme: string
 }
 
-export default class Header extends React.Component<IHeaderProps> {
-	public constructor(props: IHeaderProps) {
+export default class Header extends React.Component<{}, IHeaderState> {
+	public constructor(props: {}) {
 		super(props);
+		this.state = {
+			theme: 'dark',
+		};
+
+		this.onScroll = this.onScroll.bind(this)
 	}
+
+	public componentDidMount() {
+		window.addEventListener('scroll', this.onScroll)
+		this.onScroll();
+	}
+
+	public onScroll() {
+		const elems = Array.prototype.slice.call(document.querySelectorAll('.header-change-light, .header-change-dark'));
+		const elem = elems.find((e: HTMLElement) => {
+			const bounding = e.getBoundingClientRect();
+			return (bounding.top + bounding.height) > 60
+		});
+		if (elem) {
+			const theme = elem.classList.contains('header-change-light') ? 'light' : 'dark'
+			if (this.state.theme !== theme) {
+				this.setState({ theme })
+			}
+		}
+	}
+
 	public render() {
 		return (
-			<div className={'header header-' + this.props.theme}>
+			<div className={'header header-' + this.state.theme}>
 				<div className={'header__box'}>
 					<SearchBox />
 				</div>
